+++ common/src/main/java/com/human/common/registry/init/item/HumanPaddingBlockItems.java
@@ -141,5 +141,47 @@ public class HumanPaddingBlockItems {
                 )
         );
 
+    public static final Map<DyeColor, Supplier<BlockItem>> DYE_COLOR_TO_TILE_PADDING =
+        Collections.unmodifiableMap(
+            Arrays.stream(DyeColor.values())
+                .collect(
+                    Collectors.toMap(
+                        Function.identity(),
+                        dyeColor -> AVPBlockItems.register(
+                            dyeColor.getName() + "_tile_padding",
+                            HumanPaddingBlocks.DYE_COLOR_TO_TILE_PADDING.get(dyeColor)
+                        )
+                    )
+                )
+        );
+
+    public static final Map<DyeColor, Supplier<BlockItem>> DYE_COLOR_TO_TILE_PADDING_SLAB =
+        Collections.unmodifiableMap(
+            Arrays.stream(DyeColor.values())
+                .collect(
+                    Collectors.toMap(
+                        Function.identity(),
+                        dyeColor -> AVPBlockItems.register(
+                            dyeColor.getName() + "_tile_padding_slab",
+                            HumanPaddingBlocks.DYE_COLOR_TO_TILE_PADDING_SLAB.get(dyeColor)
+                        )
+                    )
+                )
+        );
+
+    public static final Map<DyeColor, Supplier<BlockItem>> DYE_COLOR_TO_TILE_PADDING_STAIRS =
+        Collections.unmodifiableMap(
+            Arrays.stream(DyeColor.values())
+                .collect(
+                    Collectors.toMap(
+                        Function.identity(),
+                        dyeColor -> AVPBlockItems.register(
+                            dyeColor.getName() + "_tile_padding_stairs",
+                            HumanPaddingBlocks.DYE_COLOR_TO_TILE_PADDING_STAIRS.get(dyeColor)
+                        )
+                    )
+                )
+        );
+
     public static void initialize() {}
 }